refactor generator logics
